BreakRocks and ToolBox
So with a lot of help from Lewis I managed to complete the breaking rocks function.
I also played around with making the Toolbox and confirming the switch between different tools.
The only big problem right now is that the breaking rocks function has a pretty big range. I'm not really sure how to mesaure distance yet in Unity, but this seems like it'll only be a small fix. Other than that, made a lot of progress today.
I tested all of this in "TestScene".